Flytter kode for å slå sammen forrige og nåværende perioder for innvilgelse inn i beregningService
