+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/læremidler/beregning/LæremidlerBeregningService.kt
@@ -4,6 +4,7 @@ import no.nav.tilleggsstonader.sak.behandling.domain.Saksbehandling
 import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
 import no.nav.tilleggsstonader.sak.infrastruktur.database.repository.findByIdOrThrow
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
+import no.nav.tilleggsstonader.sak.util.toYearMonth
 import no.nav.tilleggsstonader.sak.vedtak.VedtakRepository
 import no.nav.tilleggsstonader.sak.vedtak.domain.InnvilgelseEllerOpphørLæremidler
 import no.nav.tilleggsstonader.sak.vedtak.domain.StønadsperiodeBeregningsgrunnlag
@@ -12,6 +13,7 @@ import no.nav.tilleggsstonader.sak.vedtak.domain.slåSammenSammenhengende
 import no.nav.tilleggsstonader.sak.vedtak.domain.tilSortertStønadsperiodeBeregningsgrunnlag
 import no.nav.tilleggsstonader.sak.vedtak.læremidler.beregning.LæremidlerBeregnBeløpUtil.beregnBeløp
 import no.nav.tilleggsstonader.sak.vedtak.læremidler.beregning.LæremidlerBeregnUtil.grupperVedtaksperioderPerLøpendeMåned
+import no.nav.tilleggsstonader.sak.vedtak.læremidler.beregning.LæremidlerVedtaksperiodeUtil.sisteDagenILøpendeMåned
 import no.nav.tilleggsstonader.sak.vedtak.læremidler.domain.Beregningsgrunnlag
 import no.nav.tilleggsstonader.sak.vedtak.læremidler.domain.BeregningsresultatForMåned
 import no.nav.tilleggsstonader.sak.vedtak.læremidler.domain.BeregningsresultatLæremidler
@@ -36,13 +38,13 @@ class LæremidlerBeregningService(
      * Vi antar at satsen ikke endrer seg i vedtaksperioden
      * Programmet kaster feil dersom antagelsene ikke stemmer
      */
-
     fun beregn(
         behandling: Saksbehandling,
         vedtaksperioder: List<Vedtaksperiode>,
+        slåSammenMedForrigeVedtak: Boolean = true,
     ): BeregningsresultatLæremidler {
         val stønadsperioder = hentStønadsperioder(behandling.id)
-        val forrigeVedtak = hentForrigeVedtak(behandling)
+        val forrigeVedtak = if (slåSammenMedForrigeVedtak) hentForrigeVedtak(behandling) else null
 
         læremidlerVedtaksperiodeValideringService.validerVedtaksperioder(
             vedtaksperioder = vedtaksperioder,
@@ -54,12 +56,62 @@ class LæremidlerBeregningService(
         val beregningsresultatForMåned = beregnLæremidlerPerMåned(vedtaksperioder, stønadsperioder, aktiviteter)
 
         return if (forrigeVedtak != null) {
-            val revurderFra = behandling.revurderFra
-            feilHvis(revurderFra == null) { "Behandling=$behandling mangler revurderFra" }
-            BeregningsresultatLæremidler(beregningsresultatForMåned)
+            settSammenGamleOgNyePerioder(behandling, beregningsresultatForMåned, forrigeVedtak)
         } else {
             BeregningsresultatLæremidler(beregningsresultatForMåned)
         }
+    }
+
+    /**
+     * Slår sammen perioder fra forrige og nytt vedtak.
+     * Beholder perioder fra forrige vedtak frem til og med revurder-fra
+     * Bruker reberegnede perioder fra og med revurder-fra dato
+     * Dette gjøres for at vi ikke skal reberegne perioder før revurder-fra datoet
+     * Men vi trenger å reberegne perioder som løper i revurder-fra datoet då en periode kan ha endrer % eller sats
+     */
+    private fun settSammenGamleOgNyePerioder(
+        saksbehandling: Saksbehandling,
+        beregningsresultat: List<BeregningsresultatForMåned>,
+        forrigeVedtak: InnvilgelseEllerOpphørLæremidler,
+    ): BeregningsresultatLæremidler {
+        val revurderFra = saksbehandling.revurderFra
+        feilHvis(revurderFra == null) { "Behandling=$saksbehandling mangler revurderFra" }
+
+        val forrigeBeregningsresultat = forrigeVedtak.beregningsresultat
+
+        val perioderFraForrigeVedtakSomSkalBeholdes =
+            forrigeBeregningsresultat
+                .perioder
+                .filter { it.grunnlag.fom.sisteDagenILøpendeMåned() < revurderFra }
+        val nyePerioder =
+            beregningsresultat
+                .filter { it.grunnlag.fom.sisteDagenILøpendeMåned() >= revurderFra }
+
+        val nyePerioderMedKorrigertUtbetalingsdato = korrigerUtbetalingsdato(nyePerioder, forrigeBeregningsresultat)
+
+        return BeregningsresultatLæremidler(
+            perioder = perioderFraForrigeVedtakSomSkalBeholdes + nyePerioderMedKorrigertUtbetalingsdato,
+        )
+    }
+
+    private fun korrigerUtbetalingsdato(
+        nyePerioder: List<BeregningsresultatForMåned>,
+        forrigeBeregningsresultat: BeregningsresultatLæremidler,
+    ): List<BeregningsresultatForMåned> {
+        val utbetalingsdatoPerMåned =
+            forrigeBeregningsresultat
+                .perioder
+                .associate { it.grunnlag.fom.toYearMonth() to it.grunnlag.utbetalingsdato }
+
+        return nyePerioder
+            .map {
+                val utbetalingsdato = utbetalingsdatoPerMåned[it.fom.toYearMonth()]
+                if (utbetalingsdato != null) {
+                    it.medKorrigertUtbetalingsdato(utbetalingsdato)
+                } else {
+                    it
+                }
+            }
     }
 
     private fun hentForrigeVedtak(saksbehandling: Saksbehandling): InnvilgelseEllerOpphørLæremidler? =
